+++ inst/include/rxode2_control.h
@@ -137,4 +137,5 @@
 #define Rxc_indLinRichardson 132
 #define Rxc_indLinIteration 133
 #define Rxc_indLinJac 134
+#define Rxc_indLinForcing 135
 #endif // __rxode2_control_H__